fix: circular import - move funcs to utils
